implemented exception handling

+++ src/main/java/com/paathshala/dto/content/Video/VideoDetails.java
@@ -9,7 +9,7 @@ public class VideoDetails extends ContentDetails {
 
     private Integer contentLength; //in seconds
 
-    public VideoDetails(int id, String title, int courseId, String description, String contentUrl, String contentType, Byte contentSize, LocalDateTime createdAt, Integer contentLength) {
+    public VideoDetails(int id, String title, int courseId, String description, String contentUrl, String contentType, Long contentSize, LocalDateTime createdAt, Integer contentLength) {
         super(id, title, courseId, description, contentUrl, contentType, contentSize, createdAt);
         this.contentLength = contentLength;
     }
